integration-provisioning/src/main/java/org/symphonyoss/integration/provisioning/client/model/AdminApplicationWrapper.java
/**
 * Copyright 2016-2017 Symphony Integrations - Symphony LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package org.symphonyoss.integration.provisioning.client.model;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonInclude;

import java.util.ArrayList;
import java.util.List;

/**
 * Holds the information required to create or update an application on the Admin Application API.
 * Created by Bruno Campidelli on 5-jan-18.
 */
@JsonInclude(JsonInclude.Include.NON_EMPTY)
@JsonIgnoreProperties(ignoreUnknown = true)
public class AdminApplicationWrapper {

  private AdminApplicationDetailWrapper applicationInfo = null;
  private String iconUrl = null;
  private String description = null;

  public AdminApplicationDetailWrapper getApplicationInfo() {
    return applicationInfo;
  }

  public void setApplicationInfo(
      AdminApplicationDetailWrapper applicationInfo) {
    this.applicationInfo = applicationInfo;
  }

  public String getIconUrl() {
    return iconUrl;
  }

  public void setIconUrl(String iconUrl) {
    this.iconUrl = iconUrl;
  }

  public String getDescription() {
    return description;
  }

  public void setDescription(String description) {
    this.description = description;
  }
}
